+++ pepper/src/command/builtins.rs
@@ -2,7 +2,7 @@ use std::{env, path::Path, process::Stdio};
 
 use crate::{
     buffer::{BufferProperties, BufferReadError, BufferWriteError},
-    buffer_position::BufferPosition,
+    buffer_position::{BufferPosition, BufferRange},
     client::ViewAnchor,
     command::{CommandError, CommandManager, CompletionSource},
     config::{ParseConfigError, CONFIG_NAMES},
@@ -413,6 +413,183 @@ pub fn register_commands(commands: &mut CommandManager) {
             Ok(()) => Ok(()),
             Err(error) => Err(CommandError::PatternError(error)),
         }
+    });
+
+    r("buffer-list", &[], |ctx, io| {
+        io.args.assert_empty()?;
+
+        let client_handle = io.client_handle()?;
+        let buffer_view_handle = ctx
+            .editor
+            .buffer_view_handle_from_path(
+                client_handle,
+                Path::new("buffers.refs"),
+                BufferProperties::scratch(),
+                true,
+            )
+            .map_err(CommandError::BufferReadError)?;
+
+        let mut content = ctx.editor.string_pool.acquire();
+        for buffer in ctx.editor.buffers.iter() {
+            use std::fmt::Write;
+
+            let buffer_path = match buffer.path.to_str() {
+                Some(path) => path,
+                None => continue,
+            };
+
+            let _ = write!(
+                content,
+                "{} {}\n",
+                buffer_path,
+                if buffer.needs_save() { "needs save" } else { "" },
+            );
+        }
+
+        let buffer_handle = ctx
+            .editor
+            .buffer_views
+            .get(buffer_view_handle)
+            .buffer_handle;
+        let buffer = ctx.editor.buffers.get_mut(buffer_handle);
+        let range = BufferRange::between(BufferPosition::zero(), buffer.content().end());
+        buffer.delete_range(&mut ctx.editor.word_database, range, &mut ctx.editor.events);
+        buffer.insert_text(
+            &mut ctx.editor.word_database,
+            BufferPosition::zero(),
+            &content,
+            &mut ctx.editor.events,
+        );
+
+        ctx.editor.string_pool.release(content);
+
+        let client = ctx.clients.get_mut(client_handle);
+        client.set_buffer_view_handle(Some(buffer_view_handle), &ctx.editor.buffer_views);
+        Ok(())
+    });
+
+    r("lint-list", &[], |ctx, io| {
+        io.args.assert_empty()?;
+
+        let client_handle = io.client_handle()?;
+        let buffer_view_handle = ctx
+            .editor
+            .buffer_view_handle_from_path(
+                client_handle,
+                Path::new("lints.refs"),
+                BufferProperties::scratch(),
+                true,
+            )
+            .map_err(CommandError::BufferReadError)?;
+
+        let mut content = ctx.editor.string_pool.acquire();
+        for buffer in ctx.editor.buffers.iter() {
+            let buffer_path = match buffer.path.to_str() {
+                Some(path) => path,
+                None => continue,
+            };
+
+            for lint in buffer.lints.all() {
+                use std::fmt::Write;
+
+                let lint_message = lint.message(&buffer.lints);
+                let _ = write!(
+                    content,
+                    "{}:{},{}:{}\n",
+                    buffer_path,
+                    lint.range.from.line_index + 1,
+                    lint.range.from.column_byte_index + 1,
+                    lint_message
+                );
+            }
+
+            if !buffer.lints.all().is_empty() {
+                content.push('\n');
+            }
+        }
+
+        let buffer_handle = ctx
+            .editor
+            .buffer_views
+            .get(buffer_view_handle)
+            .buffer_handle;
+        let buffer = ctx.editor.buffers.get_mut(buffer_handle);
+        let range = BufferRange::between(BufferPosition::zero(), buffer.content().end());
+        buffer.delete_range(&mut ctx.editor.word_database, range, &mut ctx.editor.events);
+        buffer.insert_text(
+            &mut ctx.editor.word_database,
+            BufferPosition::zero(),
+            &content,
+            &mut ctx.editor.events,
+        );
+
+        ctx.editor.string_pool.release(content);
+
+        let client = ctx.clients.get_mut(client_handle);
+        client.set_buffer_view_handle(Some(buffer_view_handle), &ctx.editor.buffer_views);
+        Ok(())
+    });
+
+    r("breakpoint-list", &[], |ctx, io| {
+        io.args.assert_empty()?;
+
+        let client_handle = io.client_handle()?;
+        let buffer_view_handle = ctx
+            .editor
+            .buffer_view_handle_from_path(
+                client_handle,
+                Path::new("breakpoints.refs"),
+                BufferProperties::scratch(),
+                true,
+            )
+            .map_err(CommandError::BufferReadError)?;
+
+        let mut content = ctx.editor.string_pool.acquire();
+        for buffer in ctx.editor.buffers.iter() {
+            let buffer_path = match buffer.path.to_str() {
+                Some(path) => path,
+                None => continue,
+            };
+
+            for breakpoint in buffer.breakpoints() {
+                use std::fmt::Write;
+
+                let line_content =
+                    buffer.content().lines()[breakpoint.line_index as usize].as_str();
+                let _ = write!(
+                    content,
+                    "{}:{}:{}\n",
+                    buffer_path,
+                    breakpoint.line_index + 1,
+                    line_content
+                );
+            }
+
+            if !buffer.breakpoints().is_empty() {
+                content.push('\n');
+            }
+        }
+
+        let buffer_handle = ctx
+            .editor
+            .buffer_views
+            .get(buffer_view_handle)
+            .buffer_handle;
+        let buffer = ctx.editor.buffers.get_mut(buffer_handle);
+        let range = BufferRange::between(BufferPosition::zero(), buffer.content().end());
+        buffer.delete_range(&mut ctx.editor.word_database, range, &mut ctx.editor.events);
+        buffer.insert_text(
+            &mut ctx.editor.word_database,
+            BufferPosition::zero(),
+            &content,
+            &mut ctx.editor.events,
+        );
+
+        ctx.editor.string_pool.release(content);
+
+        let client = ctx.clients.get_mut(client_handle);
+        client.set_buffer_view_handle(Some(buffer_view_handle), &ctx.editor.buffer_views);
+        Ok(())
     });
 
     r("copy-command", &[], |ctx, io| {