+++ pepper/src/editor.rs
@@ -491,3 +491,4 @@ impl Editor {
         self.events.enqueue(EditorEvent::Idle);
     }
 }
+